Move validKeys to DrawableTaikoHitObject. Cleanup + reword comments.
