+++ osu.Game.Modes.Taiko/Objects/Drawable/DrawableAccentedHit.cs
@@ -17,6 +17,7 @@ namespace osu.Game.Modes.Taiko.Objects.Drawable
         private const double second_hit_window = 30;
 
         private double firstHitTime;
+        private bool firstKeyHeld;
         private Key firstHitKey;
 
         protected DrawableAccentedHit(Hit hit)
@@ -41,40 +42,44 @@ namespace osu.Game.Modes.Taiko.Objects.Drawable
                 Judgement.SecondHit = true;
         }
 
-        protected override bool OnKeyDown(InputState state, KeyDownEventArgs args)
+        protected override bool HandleKeyPress(Key key)
         {
-            // Check if we've handled the initial key
+            // Check if we've handled the first key
             if (!Judgement.Result.HasValue)
             {
-                bool result = base.OnKeyDown(state, args);
+                // First key hasn't been handled yet, attempt to handle it
+                bool handled = base.HandleKeyPress(key);
 
-                if (result)
+                if (handled)
                 {
                     firstHitTime = Time.Current;
-                    firstHitKey = args.Key;
+                    firstHitKey = key;
                 }
 
-                return result;
+                return handled;
             }
 
             // If we've already hit the second key, don't handle this object any further
             if (Judgement.SecondHit)
                 return false;
 
-            // Don't handle represses of the same key
-            if (firstHitKey == args.Key)
+            // Don't handle represses of the first key
+            if (firstHitKey == key)
                 return false;
 
             // Don't handle invalid hit key presses
-            if (!HitKeys.Contains(args.Key))
+            if (!HitKeys.Contains(key))
                 return false;
 
-            // If we're not holding the first key down still, assume the intention
-            // was not to hit the accented hit with both keys simultaneously
-            if (!state.Keyboard.Keys.Contains(firstHitKey))
-                return false;
+            // Assume the intention was to hit the accented hit with both keys only if the first key is still being held down
+            return firstKeyHeld && UpdateJudgement(true);
+        }
 
-            return UpdateJudgement(true);
+        protected override bool OnKeyDown(InputState state, KeyDownEventArgs args)
+        {
+            firstKeyHeld = state.Keyboard.Keys.Contains(firstHitKey);
+
+            return base.OnKeyDown(state, args);
         }
     }
 }
